(feat) Added debounce for volume and seek.
The SeekLeft() and SeekRight() functions were combined to one Seek()
function, and so was the VolumeUp() and VolumeDown() functions. Instead,
they now take the change as input, and add to the position/volume
accordingly. The TUI now has debounce features for both volume and seek,
since we don't want to call the functions and change position/volume
every keypress, especially if the key is held.